Parking Lot Update
The dimensions are good, but the mechanics of the damn car are still weird. It still turns very strangely, the spinning in place is not really gone. We need to think about how the rear wheels move in relation to the front wheels I think.... and how the whole car aligns. Anyway.